Revert "Simple Caching (#31)"
This reverts commit 4b428346afbbe3fda67cff8f843890f641ac5614.
