+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "ghqctoolkit"
+name = "qchub"
 version = "0.1.0"
 edition = "2024"
 
@@ -13,9 +13,9 @@ tokio = { version = "1.0", features = ["full"] }
 futures = "0.3"
 thiserror = "1.0"
 log = "0.4"
-etcetera = "0.8"
 clap = { version = "4.0", features = ["derive"], optional = true }
 anyhow = { version = "1.0", optional = true }
+etcetera = { version = "0.8", optional = true }
 clap-verbosity-flag = { version = "2.0", optional = true }
 env_logger = { version = "0.10", optional = true }
 inquire = { version = "0.7", features = ["editor"], optional = true }
@@ -24,6 +24,7 @@ inquire = { version = "0.7", features = ["editor"], optional = true }
 cli = [
     "dep:clap", 
     "dep:anyhow", 
+    "dep:etcetera",
     "dep:clap-verbosity-flag",
     "dep:env_logger",
     "dep:inquire"
@@ -35,6 +36,5 @@ tempfile = "3.0"
 insta = "1.34"
 
 [[bin]]
-name = "ghqc"
-path = "src/main.rs"
+name = "qchub"
 required-features = ["cli"]
