+++ src/git/local.rs
@@ -47,108 +47,3 @@ pub enum LocalGitError {
     #[error("Author not found for file: {0:?}")]
     AuthorNotFound(PathBuf),
 }
-
-use crate::git::GitInfo;
-
-impl LocalGitInfo for GitInfo {
-    fn commit(&self) -> Result<String, LocalGitError> {
-        let head = self.repository.head().map_err(LocalGitError::HeadError)?;
-        let commit_id = head.id().ok_or(LocalGitError::DetachedHead)?;
-        let commit_str = commit_id.to_string();
-        log::debug!("Current commit: {}", commit_str);
-        Ok(commit_str)
-    }
-
-    fn branch(&self) -> Result<String, LocalGitError> {
-        let head = self.repository.head().map_err(LocalGitError::HeadError)?;
-
-        // Try to get the branch name directly
-        if let Some(branch_name) = head.referent_name() {
-            let name_str = branch_name.as_bstr().to_string();
-            log::debug!("Raw branch reference: {}", name_str);
-
-            // Remove "refs/heads/" prefix if present
-            let final_branch = if let Some(branch) = name_str.strip_prefix("refs/heads/") {
-                branch.to_string()
-            } else {
-                name_str
-            };
-            log::debug!("Current branch: {}", final_branch);
-            Ok(final_branch)
-        } else {
-            // Fallback: we might be in detached HEAD state
-            log::debug!("No branch reference found, likely detached HEAD");
-            Ok("HEAD".to_string())
-        }
-    }
-
-    fn file_commits(&self, file: &Path) -> Result<Vec<gix::ObjectId>, LocalGitError> {
-        log::debug!("Finding commits that touched file: {:?}", file);
-        let mut commits = Vec::new();
-
-        let head_id = self
-            .repository
-            .head_id()
-            .map_err(LocalGitError::HeadIdError)?;
-
-        let revwalk = self.repository.rev_walk([head_id]);
-
-        for commit_info in revwalk.all().map_err(LocalGitError::RevWalkError)? {
-            let commit_info = commit_info.map_err(LocalGitError::TraverseError)?;
-            let commit_id = commit_info.id;
-
-            let commit = self
-                .repository
-                .find_object(commit_id)
-                .map_err(LocalGitError::FindObjectError)?
-                .try_into_commit()
-                .map_err(LocalGitError::CommitError)?;
-
-            // Check if this commit touched the file
-            if let Ok(tree) = commit.tree() {
-                let mut buffer = Vec::new();
-                if tree.lookup_entry_by_path(file, &mut buffer).is_ok() {
-                    commits.push(commit_id);
-                }
-            }
-        }
-
-        log::debug!(
-            "Found {} commits that touched file: {:?}",
-            commits.len(),
-            file
-        );
-        Ok(commits)
-    }
-
-    fn authors(&self, file: &Path) -> Result<Vec<GitAuthor>, LocalGitError> {
-        let commits = self.file_commits(file)?;
-
-        let mut res: Vec<GitAuthor> = Vec::new();
-
-        for commit_id in commits {
-            let commit = self
-                .repository
-                .find_object(commit_id)
-                .map_err(LocalGitError::FindObjectError)?
-                .try_into_commit()
-                .map_err(LocalGitError::CommitError)?;
-
-            let signature = commit.author().map_err(LocalGitError::SignatureError)?;
-            if !res.iter().any(|author| author.email == signature.email) {
-                res.push(GitAuthor {
-                    name: signature.name.to_string(),
-                    email: signature.email.to_string(),
-                });
-            }
-        }
-
-        if res.is_empty() {
-            log::warn!("No authors found for file: {:?}", file);
-            Err(LocalGitError::AuthorNotFound(file.to_path_buf()))
-        } else {
-            log::debug!("Found {} unique authors for file: {:?}", res.len(), file);
-            Ok(res)
-        }
-    }
-}
