+++ src/lib.rs
@@ -1,4 +1,3 @@
-mod cache;
 mod configuration;
 mod create;
 mod git;
@@ -7,7 +6,6 @@ mod issues;
 #[cfg(feature = "cli")]
 pub mod cli;
 
-pub use cache::DiskCache;
 pub use configuration::Configuration;
 pub use create::{MilestoneStatus, create_issue, validate_assignees};
 pub use git::{GitHubApi, GitInfo, RepoUser};
